added changes for review comments
